requirements.txt
requests==2.31.0
python-telegram-bot==20.7
beautifulsoup4==4.12.2
lxml==4.9.3
